Converted all Function interfaces, Transformable, Xform, UnmodIterable, and Or from Java to Kotlin so that we can implement the Kotlin collection interfaces as well as the Java ones.
I kept the Fn0/Fn1/Fn2/Fn3 classes in the projects, but replaced uses in the API with Kotlin's () -> T functions because Kotlin does better type inference with that.  Java users may have to to cast to (Fn1<A,B>) in order to get exception wrapping.  There's Fn1.Companion.toFn1 that will make this really brief for Java users once Kotlin 1.3 allows the JvmStatic annotation so that this method can be static-imported in Java.

Renamed all Fn0/Fn1/Fn2/Fn3.apply() to .invoke() to match Kotlin.

Unfortunately, all static methods on interfaces are currently Thing.Companion.foo() instead of Thing.foo().  This will have to be undone when Kotlin 1.3 comes out.